stage2: only default to linking system libc if linking system libs
We do need to link the system libc if linking system libraries as
they may potentially be compiled against e.g. a newer glibc version
than zig can provide. However, if not linking system libraries using
the zig provided libc is more and reliable as it does not depend on
any quirks of the host system or being able to invoke the system cc
to find include dirs.